move RobotLogListener from main to core
