+++ packages/main/pyproject.toml
@@ -36,7 +36,8 @@ packages = [
 python = "^3.6"
 docutils = "*"
 dataclasses = { version = "^0.7", python=">=3.6,<3.7" }
-rpaframework-core = "^5.3.1"
+# rpaframework-core = "^5.3.2"
+rpaframework-core = { path = "../core", develop = true}
 rpaframework-recognition = { version = "^0.6.0", optional = true }
 jsonpath-ng = "^1.5.2"
 robotframework = "^3.2.2"
